pyproject.toml
[tool.poetry]
name = "elk-upgrade"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pytz = "^2024.1"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
pandas = "2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
